robot/requirements.txt
google-generativeai==0.8.3
PyGithub==2.1.1
python-dotenv==1.0.0
